+++ examples/journald.rs
@@ -14,12 +14,10 @@
 
 #[cfg(unix)]
 fn main() {
-    use logforth::append::Journald;
+    logforth::bridge::setup_log_crate();
 
-    let append = Journald::new().unwrap();
-    logforth::builder()
-        .dispatch(|d| d.append(append))
-        .setup_log_crate();
+    let append = logforth::append::Journald::new().unwrap();
+    logforth::builder().dispatch(|d| d.append(append)).apply();
 
     log::error!("Hello, journald at ERROR!");
     log::warn!("Hello, journald at WARN!");